Replace Windows unhandled exc. with a vectored exc. handler.
