+++ DeepSkyStacker/DeepSkyStacker.vcxproj
@@ -312,6 +312,7 @@ $(QtToolsPath)\windeployqt --pdb $(TargetPath)</Command>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="ExceptionHandling.cpp" />
     <ClCompile Include="group.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">NotUsing</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">NotUsing</PrecompiledHeader>
@@ -542,6 +543,7 @@ $(QtToolsPath)\windeployqt --pdb $(TargetPath)</Command>
     <ClInclude Include="DSSVersion.h" />
     <ClInclude Include="EntropyInfo.h" />
     <QtMoc Include="ExplorerBar.h" />
+    <ClInclude Include="ExceptionHandling.h" />
     <ClInclude Include="FileProperty.h" />
     <ClInclude Include="Filters.h" />
     <ClInclude Include="FITSUtil.h" />
